+++ proxy.go
@@ -10,11 +10,11 @@ func (s *session) handlePROXY(ctx context.Context, cmd *command) context.Context
 	fields := cmd.args()
 
 	if !s.server.EnableProxyProtocol {
-		return s.reply(ctx, 550, "Proxy Protocol not enabled")
+		return s.replyEnhanced(ctx, 550, EnhancedCode{5, 7, 0}, "Proxy Protocol not enabled")
 	}
 
 	if len(fields) < 5 {
-		return s.reply(ctx, 501, "Couldn't decode the command.")
+		return s.replyEnhanced(ctx, 501, EnhancedCode{5, 5, 4}, "Couldn't decode the command.")
 	}
 
 	var (
@@ -27,12 +27,12 @@ func (s *session) handlePROXY(ctx context.Context, cmd *command) context.Context
 
 	newTCPPort, err = strconv.ParseUint(fields[3], 10, 16)
 	if err != nil {
-		return s.reply(ctx, 501, "Couldn't decode the command.")
+		return s.replyEnhanced(ctx, 501, EnhancedCode{5, 5, 4}, "Couldn't decode the command.")
 	}
 
 	tcpAddr, ok := s.peer.Addr.(*net.TCPAddr)
 	if !ok {
-		return s.reply(ctx, 502, "Unsupported network connection")
+		return s.replyEnhanced(ctx, 502, EnhancedCode{5, 5, 1}, "Unsupported network connection")
 	}
 
 	updated := &net.TCPAddr{IP: tcpAddr.IP, Port: tcpAddr.Port, Zone: tcpAddr.Zone}